basic version
Game three in a row. The meaning of the game is to use the commands to make three or more identical elements in a row.
In this game, a GAME class is implemented in which the following methods are available:
■ init fills in the field (random)
■ Move moves one crystal
■ Tick continues to tick while there is at least one change on the field
■ After each tick, visualize the field (dump ()),
■ If there are no possible movements, then mix the crystals (mix) so that new variants of movements arise - but there are no new ready triples
First, the player enters the size of the field, then the game creates the field using the init function, then the player enters a command to move the field object and if he collects three or more identical objects in a row, then these three objects disappear. Then the player again introduces new teams until he wants to leave the game by typing in the input field 'q'. If there are no new moves left on the field, then a new field is automatically generated.